Trying to fix Canvas and images.
Gtk 3 decided to get rid of GDK drawing API and replace with cairo. This
means a large portion of this code has to be rewritten.

See also:

* https://docs.gtk.org/gtk3/migrating-2to3.html#use-cairo-for-drawing
* https://wiki.gnome.org/Projects/PyGObject/IntrospectionPorting

I wonder if it would be a better idea to rewrite this in Qt or Tkinter
or pygame/SDL instead of Gtk3 or Gtk4.